requirements.txt
adal==1.2.5
aniso8601==8.1.0
appdirs==1.4.4
applicationinsights==0.11.9
astroid==2.4.2
attrs==20.3.0
azure-common==1.1.26
azure-core==1.10.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==12.0.0
azure-mgmt-storage==11.2.0
azureml-automl-core==1.20.0
azureml-contrib-services==1.20.0
azureml-core==1.20.0
azureml-dataprep==2.7.3
azureml-dataprep-native==27.0.0
azureml-dataprep-rslex==1.5.0
azureml-dataset-runtime==1.20.0
azureml-pipeline==1.20.0
azureml-pipeline-core==1.20.0
azureml-pipeline-steps==1.20.0
azureml-sdk==1.20.0
azureml-telemetry==1.20.0
azureml-train==1.20.0
azureml-train-automl-client==1.20.0
azureml-train-core==1.20.0
azureml-train-restclients-hyperdrive==1.20.0
backports.tempfile==1.0
backports.weakref==1.0.post1
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
contextlib2==0.6.0.post1
coverage==5.3.1
cryptography==3.3.1
distro==1.5.0
docker==4.4.1
dotnetcore2==2.1.20
Faker==5.5.1
flake8==3.8.4
Flask==1.1.2
Flask-RESTful==0.3.8
fusepy==3.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
inference-schema==1.1.0
iniconfig==1.1.1
install==1.3.4
isodate==0.6.0
isort==5.7.0
itsdangerous==1.1.0
jeepney==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jsonpickle==1.4.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msal==1.8.0
msal-extensions==0.2.2
msrest==0.6.19
msrestazure==0.6.4
mypy-extensions==0.4.3
ndg-httpsclient==0.5.1
numpy==1.19.5
oauthlib==3.1.0
packaging==20.8
pandas==1.2.0
pathspec==0.8.1
pluggy==0.13.1
portalocker==1.7.1
py==1.10.0
pyarrow==1.0.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rpy2==3.4.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
SecretStorage==3.3.0
six==1.15.0
snowballstemmer==2.0.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.1
zipp==3.4.0
